pyproject.toml
[tool.poetry]
name = "recotg"
version = "0.1.0"
description = "RecoTg monorepo (auth service, bot service, shared libs)"
authors = ["andreyka <you@example.com>"]
package-mode = false

# package-mode = false говорит Poetry:
# "не пытайся собирать один publishable пакет, у меня монорепа с исходниками в src/"

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.27"

# --- веб и сервисы ---
fastapi = "^0.115.0"
uvicorn = {version = "^0.30.0", extras = ["standard"]}

# --- валидация/настройки ---
pydantic = "^2.7.0"
pydantic-settings = "^2.4.0"

# --- БД (если async SQLAlchemy + Postgres) ---
SQLAlchemy = "^2.0.0"
asyncpg = "^0.29.0"
alembic = "^1.13.0"

# --- телеграм бот ---
aiogram = "^3.7.0"

# Важно: aiogram 3.x уже сидит на pydantic v2, так что мы держим всё на pydantic v2, чтобы не было конфликта,
# и НЕ ставим старые либы, которые требуют pydantic<2.

python-dotenv = "^1.0.1"
pyjwt = "^2.10.1"
psycopg2-binary = "^2.9.11"
greenlet = "^3.2.4"
pandas = "^2.3.3"
python-multipart = "^0.0.20"

# --- Группа зависимостей для ETL (Telethon + планировщик + ретраи) ---
telethon = "^1.41.2"
[tool.poetry.group.etl]
optional = true

[tool.poetry.group.etl.dependencies]
telethon = "^1.36.0"
apscheduler = "^3.10.4"
tenacity = "^8.3.0"        # экспоненциальные ретраи обёрткой
# логирование по желанию; можно включить позже
# structlog = "^24.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
httpx = "^0.27.0"
ruff = "^0.6.0"
black = "^24.0.0"

[tool.poetry.scripts]
# Потом можно будет сделать shorthands типа "poetry run auth" и "poetry run bot"
# но пока оставим пусто.
# Для ETL мы запускаем модуль напрямую: python -m services.etl.scheduler

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
